Re-organized and cleanups

+++ WolfensteinInfinite/GameObjects/ExitWall.cs
@@ -2,8 +2,8 @@
 {
     public class ExitWall : IInteractable
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
         public bool IsActivated { get; private set; } = false;
 
         public bool CanInteract(Game game) => !IsActivated;
